+++ packages/calendar-range/src/Component.test.tsx
@@ -274,8 +274,8 @@ describe('CalendarRange', () => {
 
             fireEvent.mouseEnter(days[1]);
 
-            expect(days[0]).toHaveClass('rangeStart');
-            expect(days[1]).toHaveClass('range');
+            expect(days[0].parentElement).toHaveClass('rangeStart');
+            expect(days[1].parentElement).toHaveClass('range');
         });
 
         it('should select new day, fill inputTo and end selection if clicked on same day twice', () => {
@@ -299,7 +299,7 @@ describe('CalendarRange', () => {
 
             fireEvent.mouseEnter(days[1]);
 
-            expect(days[1]).not.toHaveClass('range');
+            expect(days[1].parentElement).not.toHaveClass('range');
         });
 
         it('should start selection if clicked on same day thrice', () => {
@@ -327,7 +327,7 @@ describe('CalendarRange', () => {
 
             fireEvent.mouseEnter(days[1]);
 
-            expect(days[1]).toHaveClass('range');
+            expect(days[1].parentElement).toHaveClass('range');
         });
 
         it('should select new day, change inputFrom and inputTo values if clicked date < start date', () => {
@@ -350,7 +350,11 @@ describe('CalendarRange', () => {
             expect(inputFrom).toHaveValue(formatDate(startOfMonth(currentMonth)));
             expect(inputTo).toHaveValue(formatDate(addDays(startOfMonth(currentMonth), 2)));
 
-            expect(days[1]).toHaveClass('range');
+            fireEvent.mouseEnter(days[1]);
+
+            expect(days[0].parentElement).toHaveClass('rangeStart');
+            expect(days[1].parentElement).not.toHaveClass('rangeStart');
+            expect(days[1].parentElement).toHaveClass('range');
         });
 
         it('should select day, fill inputTo and end selection if clicked date > start date', () => {
@@ -372,7 +376,7 @@ describe('CalendarRange', () => {
 
             Array.from(days)
                 .slice(1, -1)
-                .forEach(day => expect(day).toHaveClass('range'));
+                .forEach(day => expect(day.parentElement).toHaveClass('range'));
         });
 
         it('should keep selection when month changed', async () => {
@@ -400,7 +404,7 @@ describe('CalendarRange', () => {
 
             Array.from(days)
                 .slice(0, -1)
-                .forEach(day => expect(day).toHaveClass('range'));
+                .forEach(day => expect(day.parentElement).toHaveClass('range'));
         });
     });