+++ packages/calendar/src/components/period-slider/Component.tsx
@@ -3,8 +3,6 @@ import cn from 'classnames';
 import { IconButton } from '@alfalab/core-components-icon-button';
 import { endOfWeek, startOfWeek } from 'date-fns';
 import { ChevronBackMIcon } from '@alfalab/icons-glyph/ChevronBackMIcon';
-import { SelectButton } from '../select-button';
-import { monthName } from '../../utils';
 import { formatPeriod, shiftValues } from './utils';
 
 import styles from './index.module.css';
@@ -16,11 +14,6 @@ export type PeriodSliderProps = {
      * Активная дата или период
      */
     value?: Date | [Date, Date];
-
-    /**
-     * Вид шапки — с кнопками выбора года и месяца или только период
-     */
-    view?: 'full' | 'period';
 
     /**
      * Тип периода
@@ -79,16 +72,6 @@ export type PeriodSliderProps = {
     ) => void;
 
     /**
-     * Обработчик нажатия на кнопку месяца
-     */
-    onMonthClick?: (event: MouseEvent<HTMLButtonElement>) => void;
-
-    /**
-     * Обработчик нажатия на кнопку года
-     */
-    onYearClick?: (event: MouseEvent<HTMLButtonElement>) => void;
-
-    /**
      * Идентификатор для систем автоматизированного тестирования
      */
     dataTestId?: string;
@@ -96,7 +79,6 @@ export type PeriodSliderProps = {
 
 export const PeriodSlider: FC<PeriodSliderProps> = ({
     value,
-    view = 'month-only',
     periodType = 'month',
     className,
     periodFormatter = formatPeriod,
@@ -105,8 +87,6 @@ export const PeriodSlider: FC<PeriodSliderProps> = ({
     hideDisabledArrows = false,
     onPrevArrowClick = () => null,
     onNextArrowClick = () => null,
-    onMonthClick,
-    onYearClick,
     dataTestId,
 }) => {
     const [valueFrom, valueTo] = useMemo(() => {
@@ -129,11 +109,14 @@ export const PeriodSlider: FC<PeriodSliderProps> = ({
         return [from, to];
     }, [periodType, value]);
 
-    const month = valueFrom ? monthName(valueFrom) : undefined;
-    const year = valueFrom ? valueFrom.getFullYear().toString() : undefined;
+    const showArrow = (direction: 'prev' | 'next') => {
+        if (hideDisabledArrows) {
+            const disabled = direction === 'prev' ? prevArrowDisabled : nextArrowDisabled;
+            return !disabled && valueFrom;
+        }
 
-    const showPrevButton = !(hideDisabledArrows && (prevArrowDisabled || !valueFrom));
-    const showNextButton = !(hideDisabledArrows && (nextArrowDisabled || !valueFrom));
+        return true;
+    };
 
     const handleNextArrowClick = (event: MouseEvent<HTMLButtonElement>) => {
         if (!valueFrom || !valueTo) return;
@@ -161,42 +144,15 @@ export const PeriodSlider: FC<PeriodSliderProps> = ({
         });
     };
 
-    const renderHeader = () => {
-        if (!valueFrom || !valueTo) {
-            return <span className={cn(styles.period, styles.empty)}>Укажите период</span>;
-        }
-
-        return view === 'full' ? (
-            <React.Fragment>
-                <SelectButton view='filled' className={styles.month} onClick={onMonthClick}>
-                    <span className={styles.buttonContent}>
-                        {month}
-                        <span className={styles.upDownIcon} />
-                    </span>
-                </SelectButton>
-                <SelectButton view='filled' className={styles.year} onClick={onYearClick}>
-                    <span className={styles.buttonContent}>
-                        {year}
-                        <span className={styles.upDownIcon} />
-                    </span>
-                </SelectButton>
-            </React.Fragment>
-        ) : (
-            <span className={styles.period}>{periodFormatter(valueFrom, valueTo, periodType)}</span>
-        );
-    };
-
     return (
         <div
-            className={cn(styles.component, className, {
-                [styles.full]: view === 'full',
-            })}
+            className={cn(styles.component, className)}
             aria-live='polite'
             data-test-id={dataTestId}
         >
-            {showPrevButton && (
+            {showArrow('prev') && (
                 <IconButton
-                    size="xs"
+                    size='xs'
                     className={styles.arrow}
                     icon={ChevronBackMIcon}
                     onClick={handlePrevArrowClick}
@@ -205,11 +161,17 @@ export const PeriodSlider: FC<PeriodSliderProps> = ({
                 />
             )}
 
-            {renderHeader()}
+            {valueFrom && valueTo ? (
+                <span className={styles.period}>
+                    {periodFormatter(valueFrom, valueTo, periodType)}
+                </span>
+            ) : (
+                <span className={cn(styles.period, styles.empty)}>Укажите период</span>
+            )}
 
-            {showNextButton && (
+            {showArrow('next') && (
                 <IconButton
-                    size="xs"
+                    size='xs'
                     className={styles.arrow}
                     icon={ChevronBackMIcon}
                     onClick={handleNextArrowClick}