+++ packages/code-input/src/component.screenshots.test.tsx
@@ -1,0 +1,31 @@
+import { setupScreenshotTesting, createSpriteStorybookUrl } from '../../screenshot-utils';
+
+const screenshotTesting = setupScreenshotTesting({
+    it,
+    beforeAll,
+    afterAll,
+    expect,
+});
+
+describe(
+    'CodeInput',
+    screenshotTesting({
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'CodeInput',
+                    size: { width: 200, height: 80 },
+                    knobs: {
+                        initialValues: ['', '1234'],
+                        disabled: [true, false],
+                        error: ['', 'Error'],
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
+    }),
+);