+++ packages/confirmation/src/components/index.ts
@@ -1,4 +1,2 @@
-export * from './countdown';
-export * from './sign-confirmation';
+export * from './screens';
 export * from './countdown-loader';
-export * from './overlimit';