+++ packages/confirmation/src/components/screens/index.ts
@@ -1,0 +1,5 @@
+export * from './initial';
+export * from './hint';
+export * from './fatal-error';
+export * from './temp-block';
+