+++ packages/intl-phone-input/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.14](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@6.0.13...@alfalab/core-components-intl-phone-input@6.0.14) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-intl-phone-input
+
+
+
+
+
 ## [6.0.13](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-intl-phone-input@6.0.12...@alfalab/core-components-intl-phone-input@6.0.13) (2022-02-09)
 
 **Note:** Version bump only for package @alfalab/core-components-intl-phone-input