+++ packages/password-input/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.15](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-password-input@2.1.14...@alfalab/core-components-password-input@2.1.15) (2022-02-15)
+
+**Note:** Version bump only for package @alfalab/core-components-password-input
+
+
+
+
+
 ## [2.1.14](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-password-input@2.1.13...@alfalab/core-components-password-input@2.1.14) (2022-02-09)
 
 **Note:** Version bump only for package @alfalab/core-components-password-input